GeniusGame
Project for ICS3U. In this game, a secret word will be generated to your desired length formed of characters 'r','b' and 'g'. The GUI will show you the amount of correct and incorrect guesses. Use that information to guess the secret word! Made with NetBeans and GUI Builder